pyproject.toml
[project]
name = "todoist-sh"
version = "0.1.0"
description = "todoist CLI interface"
readme = "README.md"
requires-python = ">=3.13"
authors = [{name = "Darrion Burgess", email = "darrionburgess@gmail.com", github="ca-mantis-shrimp"}]

dependencies = ["tomlkit", "todoist-api-python", "python-dotenv", "keyring", "keyrings.cryptfile", "requests", "platformdirs"]

[project.optional-dependencies]
test = ["pytest"]

[project.scripts]
tdst = "todoist_sh:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
